`janus_cli`: add dry-run mode
Building on the previously introduced dry-run support in `Datastore`,
`janus_cli` can now be run with a `--dry-run` flag. In this mode, no
real modifications to the datastore (in the case of `write-schema` or
`provision-tasks`) or Kubernetes secrets (in the case of
`create-datastore-key`) are made.

We also refactor some of the command line option and configuration file
handling to reduce code duplication in `Command::execute`.

Resolves #529